My Capstone Project
This took my 2 mouths to do, looking back on it now there are lots of
things I would change and if I continue working on it I will. I was
under a lot of pressure to get it in time and I jimmy rigged a lot but I
got it Done. Even if I would do the hit detection from scratch I did all
the coding and it does what it needs to. A majority I code in C++ like
hit detection, AI and health displays other things like animation states
were done in blueprints. Please tell me what you think or if there is
anything I should change and I will add it to me ChangeLog.